latest version. 各种类型比对分别单独使用阻塞队列，避免相互之间干扰

+++ src/main/java/HAFPIS/service/OneToF_FPTT.java
@@ -151,9 +151,9 @@ public class OneToF_FPTT extends Recog implements Runnable {
                             tempScore = temp;
                         }
                         if (j >= 0 && j < 10) {
-                            fpttRec.rpscores[j] = temp;
+                            fpttRec.rpscores[j] = (int) temp;
                         } else {
-                            fpttRec.fpscores[j - 10] = temp;
+                            fpttRec.fpscores[j - 10] = (int) temp;
                         }
                     }
                     fpttRec.score = tempScore;
